% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hexbin_label.R
\name{make_hexbin_label}
\alias{make_hexbin_label}
\title{Group label plot position.}
\usage{
make_hexbin_label(sce, col)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}}
or \code{\link[Seurat]{Seurat-class}} object.}

\item{col}{The name referring to one column in meta data for which
the label position on the plot is calculated for every level. The chosen
column needs to be a factor.}
}
\value{
A dataframe.
}
\description{
Group label plot position.
}
\examples{
#' # For Seurat object
library(Seurat)
data("pbmc_small")
pbmc_small <- make_hexbin(pbmc_small, 10, dimension_reduction = "PCA")
make_hexbin_label(pbmc_small, col = "RNA_snn_res.1")
}
