% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hexbin_density_shiny.R
\name{plot_hexbin_density_shiny}
\alias{plot_hexbin_density_shiny}
\title{Plot of density of observations from single cell data
   in bivariate hexagon cells as shiny instance.}
\usage{
plot_hexbin_density_shiny(sce, min_nbins, max_nbins, dimension_reduction)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}}
or \code{\link[Seurat]{Seurat-class}} object.}

\item{min_nbins}{The miniumum number of bins partitioning the range of the 
first component of the chosen dimension reduction.}

\item{max_nbins}{The miniumum number of bins partitioning the range of the 
first component of the chosen dimension reduction.}

\item{dimension_reduction}{A string indicating the reduced dimension
result to calculate hexagon cell representation of.}
}
\value{
An object that represents the app.
}
\description{
Plot of density of observations from single cell data
   in bivariate hexagon cells as shiny instance.
}
\details{
This function opens a shiny instance, which allows to investigate 
   the effect of the resolution parameter. The user can change the resolution
   using the slider.
}
\examples{
# For Seurat object
\dontrun{
library(Seurat)
data("pbmc_small")
plot_hexbin_density_shiny(pbmc_small,3, 10, dimension_reduction = "PCA")
}
}
\seealso{
\code{\link{plot_hexbin_density}}
}
