% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hexbin_feature_shiny.R
\name{plot_hexbin_feature_shiny}
\alias{plot_hexbin_feature_shiny}
\title{Plot of feature expression of single cell data in bivariate hexagon cells as
    shiny instance.}
\usage{
plot_hexbin_feature_shiny(sce, mod = "RNA", type, feature, action,
  min_nbins, max_nbins, dimension_reduction)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}}
or \code{\link[Seurat]{Seurat-class}} object.}

\item{mod}{A string referring to the name of the modality used for plotting.
For RNA modality use "RNA". For other modalities use name of alternative 
object for the \code{\link[SingleCellExperiment]{SingleCellExperiment}} 
or the name of the assay for the \code{\link[Seurat]{Seurat-class}} 
object.}

\item{type}{A string referring to the type of assay in the
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object or the
data transformation in the \code{\link[Seurat]{Seurat-class}} object.}

\item{feature}{A string referring to the name of one feature.}

\item{action}{A strings pecifying how meta data of observations in
binned  hexagon cells are to be summarized. Possible actions are
\code{prop_0}, \code{mode}, \code{mean} and \code{median} (see 
\code{\link{plot_hexbin_feature}}).}

\item{min_nbins}{The miniumum number of bins partitioning the range of the 
first component of the chosen dimension reduction.}

\item{max_nbins}{The miniumum number of bins partitioning the range of the 
first component of the chosen dimension reduction.}

\item{dimension_reduction}{A string indicating the reduced dimension
result to calculate hexagon cell representation of.}
}
\value{
An object that represents the app.
}
\description{
Plot of feature expression of single cell data in bivariate hexagon cells as
    shiny instance.
}
\details{
This function opens a shiny instance, which allows to investigate 
   the effect of the resolution parameter. The user can change the resolution
   using the slider. Each hexagon is clickable, which will plot the 
   observations in the chosen hexagons in a histograms below.
}
\examples{
# For Seurat object
\dontrun{
library(Seurat)
data("pbmc_small")
plot_hexbin_feature_shiny(pbmc_small, type="counts", feature="TALDO1", 
   action="prop_0", min_nbins=2, max_nbins=10, dimension_reduction="PCA",
   mod="RNA")
}
}
\seealso{
\code{\link{plot_hexbin_feature}}
}
