% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hexbin_gene.R
\docType{methods}
\name{plot_hexbin_gene}
\alias{plot_hexbin_gene}
\alias{plot_hexbin_gene,SingleCellExperiment-method}
\alias{plot_hexbin_gene,Seurat-method}
\title{Plot of gene expression of single cells in bivariate hexagon cells.}
\usage{
plot_hexbin_gene(sce, type, gene, action, title = NULL, xlab = NULL,
  ylab = NULL)

\S4method{plot_hexbin_gene}{SingleCellExperiment}(sce, type, gene, action,
  title = NULL, xlab = NULL, ylab = NULL)

\S4method{plot_hexbin_gene}{Seurat}(sce, type, gene, action,
  title = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}}
or \code{\link[Seurat]{Seurat}} object.}

\item{type}{A string referring to the type of assay in the
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object or the
data transformation in the \code{\link[Seurat]{Seurat}} object.}

\item{gene}{A string referring to the name of one gene.}

\item{action}{A strings pecifying how gene expression of observations in
binned hexagon cells are to be summarized. Possible actions are
\code{prop_0}, \code{mode}, \code{mean} and \code{median} (see details).}

\item{title}{A string containing the title of the plot.}

\item{xlab}{A string containing the title of the x axis.}

\item{ylab}{A string containing the title of the y axis.}
}
\value{
A \code{\link{ggplot2}{ggplot}} object.
}
\description{
Plot of gene expression of single cells in bivariate hexagon cells.
}
\details{
This function plots the expression of any gene in the hexagon cell
  representation calculated with \code{\link{make_hexbin}}. The chosen gene
  expression is summarized by one of four actions \code{prop_0}, \code{mode},
  \code{mean} and \code{median}:

  \describe{
    \item{\code{prop_0}}{Returns the proportion of observations in the bin
     greater than 0. The associated meta data column needs to be numeric.}
    \item{\code{mode}}{Returns the mode of the observations in the bin. The
     associated meta data column needs to be numeric.}
    \item{\code{mean}}{Returns the mean of the observations in the bin. The
     associated meta data column needs to be numeric.}
     \item{\code{median}}{Returns the median of the observations in the bin.
     The associated meta data column needs to be numeric.}
  }
}
\section{Methods (by class)}{
\itemize{
\item \code{SingleCellExperiment}: Plot of gene expression into hexagon cell for
SingleCellExperiment object.

\item \code{Seurat}: Plot of gene expression into hexagon cell for
Seurat object.
}}

\examples{
# For Seurat object
library(Seurat)
data("pbmc_small")
pbmc_small <- make_hexbin(pbmc_small, 10, dimension_reduction = "PCA")
plot_hexbin_gene(pbmc_small, type="counts", gene="TALDO1", action="prop_0")
# For SingleCellExperiment object
\dontrun{
library(TENxPBMCData)
library(scater)
tenx_pbmc3k <- TENxPBMCData(dataset = "pbmc3k")
rm_ind <- calcAverage(tenx_pbmc3k)<0.1
tenx_pbmc3k <- tenx_pbmc3k[!rm_ind,]
colData(tenx_pbmc3k) <- cbind(colData(tenx_pbmc3k),
     perCellQCMetrics(tenx_pbmc3k))
tenx_pbmc3k <- normalize(tenx_pbmc3k)
tenx_pbmc3k <- runPCA(tenx_pbmc3k)
tenx_pbmc3k <- make_hexbin( tenx_pbmc3k, 20, dimension_reduction = "PCA")
plot_hexbin_gene(tenx_pbmc3k, type="logcounts",
gene="ENSG00000135250", action="mean")
plot_hexbin_gene(tenx_pbmc3k, type="logcounts",
gene="ENSG00000135250", action="mode")
}
}
