% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hexbin_interact.R
\docType{methods}
\name{plot_hexbin_interact}
\alias{plot_hexbin_interact}
\alias{plot_hexbin_interact,SingleCellExperiment-method}
\alias{plot_hexbin_interact,Seurat-method}
\title{Plot of interaction of expression of single cells in bivariate hexagon cells.}
\usage{
plot_hexbin_interact(sce, mod, type, feature, interact, title = NULL,
  xlab = NULL, ylab = NULL)

\S4method{plot_hexbin_interact}{SingleCellExperiment}(sce, mod, type,
  feature, interact, title = NULL, xlab = NULL, ylab = NULL)

\S4method{plot_hexbin_interact}{Seurat}(sce, mod, type, feature, interact,
  title = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}}
or \code{\link[Seurat]{Seurat-class}} object.}

\item{mod}{A vector of strings referring to the names of the modularities.
For \code{\link[SingleCellExperiment]{SingleCellExperiment}} use "RNA" to
access the RNA expression data stored as the main experiment type.}

\item{type}{A vector of strings referring to the types of assays in the
\code{\link[SingleCellExperiment]{SingleCellExperiment}} or the types of
transformation in \code{\link[Seurat]{Seurat-class}} object.}

\item{feature}{A vector of strings referring to the names of one features in
the same order as the vector of modularities.}

\item{interact}{A string specifying how interaction between features is
calculated. Possible interaction measures are
\code{corr_spearman} and \code{mi} (see details).}

\item{title}{A string containing the title of the plot.}

\item{xlab}{A string containing the title of the x axis.}

\item{ylab}{A string containing the title of the y axis.}
}
\value{
A \code{\link{ggplot2}{ggplot}} object.
}
\description{
Plot of interaction of expression of single cells in bivariate hexagon cells.
}
\details{
This function plots the interaction between any features in the
   hexagon cell representation calculated with \code{\link{make_hexbin}}. The
   interaction between the chosen features is calculated by one of two
   measurers \code{corr_spearman}, and \code{mi}:

   \describe{
      \item{\code{mi}}{Returns the mutual information coefficient.}
      \item{\code{corr_spearman}}{Returns the Spearman correlation.}
   }
}
\section{Methods (by class)}{
\itemize{
\item \code{SingleCellExperiment}: Plot of gene expression into hexagon cell
for SingleCellExperiment object.

\item \code{Seurat}: Plot of gene expression into hexagon cell
for Seurat object.
}}

\examples{
# For Seurat object
library(Seurat)
data("pbmc_small")
pbmc_small <- make_hexbin(pbmc_small, 10, dimension_reduction = "PCA")
protein <- matrix(rnorm(10 * ncol(pbmc_small)), ncol = ncol(pbmc_small))
rownames(protein) <- paste0("A", seq(1, 10, 1))
colnames(protein) <- colnames(pbmc_small)
pbmc_small[["ADT"]] <- CreateAssayObject(counts = protein)
plot_hexbin_interact(pbmc_small,
  type = c("counts", "counts"),
  mod = c("RNA", "ADT"), feature = c("CD7", "A1"), interact = "mi"
)
}
