% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schex.R
\docType{package}
\name{schex}
\alias{schex}
\alias{schex-package}
\title{schex: A package for plotting hexbin plots for single cell omics data.}
\description{
The schex package provides three categories of two functions types:
make_hexbin and plot_hexbin functions.
}
