% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hexbin_density.R
\docType{methods}
\name{plot_hexbin_density}
\alias{plot_hexbin_density}
\alias{plot_hexbin_density,SingleCellExperiment-method}
\alias{plot_hexbin_density,Seurat-method}
\title{Plot of density of observations from single cell data
   in bivariate hexagon cells.}
\usage{
plot_hexbin_density(sce, title = NULL, xlab = NULL, ylab = NULL)

\S4method{plot_hexbin_density}{SingleCellExperiment}(sce, title = NULL,
  xlab = NULL, ylab = NULL)

\S4method{plot_hexbin_density}{Seurat}(sce, title = NULL, xlab = NULL,
  ylab = NULL)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}}
or \code{\link[Seurat]{Seurat}} object.}

\item{title}{A string containing the title of the plot.}

\item{xlab}{A string containing the title of the x axis.}

\item{ylab}{A string containing the title of the y axis.}
}
\value{
A \code{\link{ggplot2}{ggplot}} object.
}
\description{
Plot of density of observations from single cell data
   in bivariate hexagon cells.
}
\section{Methods (by class)}{
\itemize{
\item \code{SingleCellExperiment}: Plot of cell density in hexagon cell for
SingleCellExperiment object.

\item \code{Seurat}: Plot of cell density in hexagon cell for
Seurat object.
}}

\examples{
#' # For Seurat object
library(Seurat)
data("pbmc_small")
pbmc_small <- make_hexbin(pbmc_small, 10, dimension_reduction = "PCA")
plot_hexbin_density(pbmc_small)
\dontrun{
library(TENxPBMCData)
library(scater)
tenx_pbmc3k <- TENxPBMCData(dataset = "pbmc3k")
rm_ind <- calcAverage(tenx_pbmc3k)<0.1
tenx_pbmc3k <- tenx_pbmc3k[!rm_ind,]
tenx_pbmc3k <- calculateQCMetrics(tenx_pbmc3k)
tenx_pbmc3k <- normalize(tenx_pbmc3k)
tenx_pbmc3k <- runPCA(tenx_pbmc3k)
tenx_pbmc3k <- make_hexbin( tenx_pbmc3k, 20, dimension_reduction = "PCA")
plot_hexbin_density(tenx_pbmc3k)
}
}
