% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hexbin_feature.R
\docType{methods}
\name{plot_hexbin_feature}
\alias{plot_hexbin_feature}
\alias{plot_hexbin_feature,SingleCellExperiment-method}
\alias{plot_hexbin_feature,Seurat-method}
\title{Plot of external feature expression of single cells in bivariate hexagon
   cells.}
\usage{
plot_hexbin_feature(sce, mod, type, feature, action, title = NULL,
  xlab = NULL, ylab = NULL)

\S4method{plot_hexbin_feature}{SingleCellExperiment}(sce, mod, type,
  feature, action, title = NULL, xlab = NULL, ylab = NULL)

\S4method{plot_hexbin_feature}{Seurat}(sce, mod, type, feature, action,
  title = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}}
or \code{\link[Seurat]{Seurat-class}} object.}

\item{mod}{A string referring to the name of the alternative object in a
\code{\link[SingleCellExperiment]{SingleCellExperiment}} or the assay in a
\code{\link[Seurat]{Seurat-class}} object that stores the protein information.}

\item{type}{A string referring to the type of assay in the
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object or the
data transformation in the \code{\link[Seurat]{Seurat-class}} object.}

\item{feature}{A string referring to the name of one external feature.}

\item{action}{A strings pecifying how meta data of observations in
binned  hexagon cells are to be summarized. Possible actions are
\code{prop_0}, \code{mode}, \code{mean} and \code{median} (see details).}

\item{title}{A string containing the title of the plot.}

\item{xlab}{A string containing the title of the x axis.}

\item{ylab}{A string containing the title of the y axis.}
}
\value{
A \code{\link{ggplot2}{ggplot}} object.
}
\description{
Plot of external feature expression of single cells in bivariate hexagon
   cells.
}
\details{
This function plots the expression of any feature in the hexagon
   cell representation calculated with \code{\link{make_hexbin}}. The chosen
   gene expression is summarized by one of four actions \code{prop_0},
   \code{mode}, \code{mean} and \code{median}:

   \describe{
     \item{\code{prop_0}}{Returns the proportion of observations in the bin
      greater than 0. The associated meta data column needs to be numeric.}
     \item{\code{mode}}{Returns the mode of the observations in the bin. The
      associated meta data column needs to be numeric.}
     \item{\code{mean}}{Returns the mean of the observations in the bin. The
      associated meta data column needs to be numeric.}
      \item{\code{median}}{Returns the median of the observations in the bin.
      The associated meta data column needs to be numeric.}
   }
}
\section{Methods (by class)}{
\itemize{
\item \code{SingleCellExperiment}: Plot of gene expression into hexagon
cell for SingleCellExperiment object.

\item \code{Seurat}: Plot of gene expression into hexagon cell
for Seurat object.
}}

\examples{
# For Seurat object
library(Seurat)
data("pbmc_small")
pbmc_small <- make_hexbin(pbmc_small, 10, dimension_reduction = "PCA")
protein <- matrix(rnorm(10* ncol(pbmc_small)), ncol=ncol(pbmc_small))
rownames(protein) <- paste0("A", seq(1,10,1))
colnames(protein) <- colnames(pbmc_small)
pbmc_small[["ADT"]] <- CreateAssayObject(counts = protein)
pbmc_small <- make_hexbin(pbmc_small, 10, dimension_reduction = "PCA")
plot_hexbin_feature(pbmc_small, type="counts", mod="ADT",
    feature="A1", action="prop_0")
}
