% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schex.R
\docType{package}
\name{schex}
\alias{schex}
\title{schex: A package for plotting hexbin plots for single cell omics data.}
\description{
Builds hexbin plots for variables and dimension reduction stored
single cell omics data such as SingleCellExperiment. The
ideas used in this package are based on the excellent work of Dan Carr,
Nicholas Lewin-Koh, Martin Maechler and Thomas Lumley.
}
\details{
Please see the help pages listed below:
   \itemize{
      \item{\code{\link{make_hexbin}}}
      \item{\code{\link{plot_hexbin_feature}}}
      \item{\code{\link{plot_hexbin_meta}}}
   }

 Also see the vignettes for more usage examples.

 Please report issues and suggest improvements at Github:

 \url{https://github.com/SaskiaFreytag/schex}
}
